refactor file locations
