--- app/src/main/java/com/example/fridgerec/interfaces/LithoUIChangeHandler.java
@@ -1,12 +1,0 @@
-package com.example.fridgerec.interfaces;
-
-import androidx.lifecycle.ViewModel;
-
-import com.example.fridgerec.model.EntryItem;
-import com.example.fridgerec.model.EntryItemList;
-
-import java.util.List;
-
-public interface LithoUIChangeHandler {
-
-}
