+++ app/src/main/java/com/example/fridgerec/litho/FoodGroupsSectionSpec.java
@@ -1,5 +1,6 @@
-package com.example.fridgerec.activities.lithoSpecs;
+package com.example.fridgerec.litho;
 
+import com.example.fridgerec.activities.lithoSpecs.ListSection;
 import com.example.fridgerec.model.EntryItem;
 import com.facebook.litho.annotations.Prop;
 import com.facebook.litho.sections.Children;
